pyproject.toml
[project]
name = "outfit-retargeting-system"
version = "2.0.0"
description = "衣装リターゲットシステム - Blender上で素体に衣装を自動フィッティング"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "GPL-3.0-only" }
dependencies = ["numpy", "scipy"]

[dependency-groups]
dev = ["ruff", "fake-bpy-module-latest"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W", "B", "UP", "C4", "SIM", "RUF"]
ignore = [
    "E501",
    "B023",
    "RUF001",
    "RUF013",
    "RUF059",
    "SIM102",
    "SIM103",
    "SIM105",
    "SIM108",
]

[tool.ruff.format]
quote-style = "double"
